🔧 refactor: 더미객체 삭제

+++ finafanAndroid/app/src/main/java/com/a702/finafan/data/savings/dto/request/SavingCreateRequest.kt
@@ -1,9 +1,10 @@
 package com.a702.finafan.data.savings.dto.request
 
+import com.a702.finafan.domain.savings.model.Account
 import com.a702.finafan.domain.savings.model.SavingCreate
 
 data class SavingCreateRequest(
-    val entertainerId: Int,
+    val entertainerId: Long,
     val accountName: String,
     val connectAccount: Account
 )
